+++ docs/en/changes/changes.md
@@ -58,6 +58,7 @@
 * Add new HTTP URI grouping engine with benchmark. 
 * [Breaking Change] Use the new HTTP URI grouping engine to replace the old regex based mechanism.
 * support `sumLabeled` in `MAL`.
+* Envoy ALS generated relation metrics considers http status codes >= 400 has an error at the client side.
 
 #### UI
 * Revert: cpm5d function. This feature is cancelled from backend.